+++ src/GiViewHelper.mm
@@ -578,7 +578,7 @@ static GiViewHelper *_sharedInstance = nil;
     return [_view coreView]->zoomToExtent();
 }
 
-- (BOOL)zoomToExtent:(float)margin {
+- (BOOL)zoomToExtent:(CGFloat)margin {
     return [_view coreView]->zoomToExtent(margin);
 }
